Fix for Uploader.php to prevent uploads failing
The ternary operator has a bracket in the wrong place, which can cause uploads to fail.

For example, instead of "/tmp/cldupload.7Ua4gD.avi", I only get ".avi". This is because the incorrect bracket placement turns the entire first half of the string into the condition for the ternary operator, which evaluates to true, meaning the true part of the operator is used, which is just the extension.

Since the path is stripped completely, the temporary file is created in the current working directory, which the script may not have write access to. This causes a "Server returned unexpected status code - 0 - ", as the file chunk is not written, and can then not be read back for upload.

My change moves one bracket to encompass the entire ternary operator, which causes it to evaluate correctly, eliminating the issue.